requirements.txt
# Core libraries
numpy==1.26.4
pillow==10.3.0
tqdm==4.66.2
opencv-python==4.9.0.80

# PyTorch ecosystem (latest stable for YOLOv8 and general deep learning)
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1

# YOLOv8
ultralytics==8.3.94

# Labeling tools
labelme==5.7.0
# labelImg==1.8.6  # (commented out due to PyQt5 issues)

# DeepLab (Segmentation models)
segmentation-models-pytorch==0.3.3
timm==0.9.2

# Evaluation dependencies
pycocotools>=2.0.6
matplotlib>=3.7.0

# Data augmentation for Faster R-CNN
albumentations==1.4.0
